requirements.txt
django
djangorestframework
numpy
scikit-learn
joblib
